backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.0
pydantic[email]==2.9.0
python-multipart==0.0.9
anthropic==0.39.0
openai==1.55.0
google-generativeai==0.8.0
Pillow==10.4.0
aiofiles==24.1.0
